refactor(core): move challenge-response primitives into treeship-core (slice 3a)
Presentation verification starts moving into core, mirroring resolution.

- New treeship_core::verify::presentation with challenge_canonical and
  check_challenge, lifted verbatim from packages/cli/src/commands/present.rs.
  challenge_canonical is byte-critical (a single-byte change to its domain
  separation would break every previously signed challenge), so it moves
  unchanged; both challenge tests move with it and pass in core.
- present.rs imports them from core; the pub(crate) copies and their tests are
  deleted, and the now-dead HashMap / Signature / sha2 imports removed.

No behavior change. verify_staple (needs `now` injected) and a pure
verify_presentation orchestrator + WASM/SDK exposure are the next steps.
145 CLI tests + 62 core verify tests pass; build + fmt clean.
